Refactor contact resolver
