+++ src/app/contacts/module.js
@@ -1,8 +1,8 @@
 import angular from 'angular';
 import angularMessages from 'angular-messages';
 import angularResource from 'angular-resource';
+import contactFactory from './services/contact_factory';
 import contactFormComponent from './components/contact_form_component';
-import contactResource from './services/contact_resource';
 import states from './states';
 import uiRouter from 'angular-ui-router';
 
@@ -11,6 +11,6 @@ export default angular.module('app.contacts', [
   angularResource,
   uiRouter
 ])
-  .factory('Contact', contactResource)
+  .factory('Contact', contactFactory)
   .component('contactForm', contactFormComponent)
   .config(states);
