+++ src/app/contacts/services/contact_factory.js
@@ -14,10 +14,3 @@ export default function($resource) {
     delete: { method: 'DELETE' }
   });
 }
-
-export function resolveContact($stateParams, Contact) {
-  'ngInject';
-
-  const { id } = $stateParams;
-  return Contact.get({ id }).$promise;
-}
