+++ src/app/contacts/services/contact_factory_spec.js
