+++ src/app/contacts/services/resolve_contact.js
@@ -1,0 +1,6 @@
+export default function($stateParams, Contact) {
+  'ngInject';
+
+  const { id } = $stateParams;
+  return Contact.get({ id }).$promise;
+}
